Rl 64bit nonce

+++ src/primitives/block.h
@@ -28,7 +28,7 @@ public:
     uint32_t nBits;
     uint32_t nHeight;
     uint256 hashMix;
-    uint32_t nNonce;
+    uint64_t nNonce;
 
     CBlockHeader()
     {
